Update whitelist.yaml
Whitelist sites accessed by 4everland.io gateway.
As a provider of a public gateway for IPFS contents, 4everland.io has served many small and medium sized projects with IPFS pinning/gateway services. Like the PR of ICP gateway, some contracts associated with these sites have been flagged as malicious and subsequently blacklisted. However, it is important to note that this does not imply that the entire gateway and all contracts on the 4everland.io are harmful. In light of the numerous users and projects that rely on our gateway for accessing hosted websites, it is advisable to maintain the whitelist and keep the gateway accessible.